refactor assets without any changes in functionality, pure organizational
