+++ apps/builder/app/builder/shared/assets/delete-assets.ts
@@ -1,0 +1,11 @@
+import type { Asset } from "@webstudio-is/sdk";
+import { $assets } from "~/shared/nano-states";
+import { serverSyncStore } from "~/shared/sync/sync-stores";
+
+export const deleteAssets = (assetIds: Asset["id"][]) => {
+  serverSyncStore.createTransaction([$assets], (assets) => {
+    for (const assetId of assetIds) {
+      assets.delete(assetId);
+    }
+  });
+};
