+++ apps/builder/app/builder/shared/assets/index.ts
@@ -1,4 +1,6 @@
-export * from "./use-assets";
+export { useAssets } from "./use-assets";
+export * from "./delete-assets";
+export { uploadAssets } from "./upload-assets";
 export * from "./types";
 export * from "./separator";
 export * from "./assets-shell";
